Package as self-contained Windows executable
Add a build.py that uses pyinstaller to create a self-contained Windows
exe for easier distribution. This allows users to run the application
without having to also have a development environment ready.

Add a Github Actions workflow to perform this build & optionally to
create releases on the repository based on the result.