Fix release missing zip; remove old build.dat
Add back the --zip flag to the build step in the release execution to
create the zipfile that will be attached to the release.

The upload-artifact action _must_ accept a directory; if passed a zip
file it will create a nested zipfile output containing it.

Releases on the other hand can accept a zipfile/binary artifact, so we
create one ourselves and pass it.